fix(route information popup): make it so the text in route info doesn't split mid-word
Previously divs and floating features were used for the layout of the popups of route information.
Though this provided suitable display normally, the size of the leaflet popup is based on the size
of the div. When the icon for the profile is floated, this moves the text but doesn't change the
size of the underlying div, which meant that the text got wrapped into new lines. In this commit,
the divs have been replaced with spans using inline-block display so that the parent div resizes as
expected.

fix #277